requirements.txt
typer
ruff
rich
pypdf
google-genai
python-dotenv
